rename torch to avoid import confusion
